+++ packages/jest-runtime/src/index.js
@@ -26,6 +26,7 @@ import {escapePathForRegex} from 'jest-regex-util';
 import Snapshot from 'jest-snapshot';
 import fs from 'graceful-fs';
 import stripBOM from 'strip-bom';
+import {sync as glob} from 'glob';
 import ScriptTransformer from './script_transformer';
 import shouldInstrument from './should_instrument';
 import {run as cliRun} from './cli';
@@ -412,10 +413,48 @@ class Runtime {
   }
 
   requireModuleOrMock(from: Path, moduleName: string) {
-    if (this._shouldMock(from, moduleName)) {
-      return this.requireMock(from, moduleName);
-    } else {
-      return this.requireModule(from, moduleName);
+    try {
+      if (this._shouldMock(from, moduleName)) {
+        return this.requireMock(from, moduleName);
+      } else {
+        return this.requireModule(from, moduleName);
+      }
+    } catch (e) {
+      if (e.code === 'MODULE_NOT_FOUND') {
+        if (!path.isAbsolute(moduleName) && path.extname(moduleName) === '') {
+          const dirname = path.dirname(from);
+          const pathToModule = path.resolve(dirname, moduleName);
+
+          try {
+            const matches = glob(`${pathToModule}.*`)
+              .map(match => {
+                const relativePath = path.relative(dirname, match);
+
+                return path.dirname(match) === dirname
+                  ? `./${relativePath}`
+                  : relativePath;
+              })
+              .map(match => `\t'${match}'`)
+              .join('\n');
+
+            if (matches) {
+              const foundMessage = `\n\nHowever, Jest was able to find:\n${matches}`;
+
+              const moduleFileExtensions = this._config.moduleFileExtensions
+                .map(ext => `'${ext}'`)
+                .join(', ');
+
+              const appendedMessage =
+                foundMessage +
+                "\n\nYou might want to include a file extension in your import, or update your 'moduleFileExtensions', which is currently " +
+                `[${moduleFileExtensions}].\n\nSee https://jestjs.io/docs/en/configuration#modulefileextensions-array-string`;
+
+              e.message += appendedMessage;
+            }
+          } catch (ignored) {}
+        }
+      }
+      throw e;
     }
   }